# server-aplikasi-bangkit/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Authentication API",
        "description": "API for user authentication and registration",
        "version": "1.0.0"
    },
    "paths": {
...
        "/operator/rute/{id}": {
            "get": {
                "tags": [
                    "Operator Rute"
                ],
                "summary": "Dapatkan detail rute",
                "operationId": "getRuteDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Rute",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rute"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rute tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Operator Rute"
                ],
                "summary": "Update rute",
                "operationId": "updateRute",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Rute",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_rute": {
                                        "type": "string",
                                        "example": "Rute B Updated"
                                    },
                                    "titik_awal": {
                                        "type": "string",
                                        "example": "Stasiun Utama"
                                    },
                                    "koordinat_awal": {
                                        "type": "array",
                                        "items": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 106.1112
                                        }
                                    },
                                    "titik_akhir": {
                                        "type": "string",
                                        "example": "Bandara Internasional"
                                    },
                                    "koordinat_akhir": {
                                        "type": "array",
                                        "items": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 106.9998
                                        }
                                    },
                                    "path": {
                                        "type": "array",
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "type": "number",
                                                "format": "float"
                                            }
                                        },
                                        "example": [
                                            [
                                                106.1111,
                                                -6.1111
                                            ],
                                            [
                                                106.2223,
                                                -6.2223
                                            ],
                                            [
                                                106.9999,
                                                -6.9999
                                            ]
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rute updated successfully"
                                        },
                                        "rute": {
                                            "$ref": "#/components/schemas/Rute"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rute tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Operator Rute"
                ],
                "summary": "Hapus rute",
                "operationId": "deleteRute",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Rute",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rute deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rute tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Point": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "Point"
                    },
                    "coordinates": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float",
                            "example": 106.1234
                        }
                    }
                },
                "type": "object"
            },
            "LineString": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "LineString"
                    },
                    "coordinates": {
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {
                                "type": "number",
                                "format": "float"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "Rute": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_rute": {
                        "type": "string",
                        "example": "Rute A"
                    },
                    "titik_awal": {
                        "type": "string",
                        "example": "Pasar"
                    },
                    "koordinat_awal": {
                        "$ref": "#/components/schemas/Point"
                    },
                    "titik_akhir": {
                        "type": "string",
                        "example": "Terminal"
                    },
                    "koordinat_akhir": {
                        "$ref": "#/components/schemas/Point"
                    },
                    "path": {
                        "$ref": "#/components/schemas/LineString"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}